pyproject.toml
[tool.poetry]
name = "multiburst"
version = "0.1.0"
description = ""
authors = ["Gowthamaraj rajendran <gowthamaraj360@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
jinja2 = "^3.1.2"
pyyaml = "^6.0.1"
envparse = "^0.2.0"
ansible = "^8.4.0"
ansible-lint = "6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
